Replace enumerator in GameObject.Update
If the collection of objects is for some reason modified (particle
effects for instance) then the loop explodes. By copying a list of
object keys into a seperate collection (that is correct at that moment)
we ensure that we can continue iterating over that list of keys
regardless of new objects being added to the list.