Fix memory corruption on server restart
NetServer.Stop would cause the LuaState that has just been created in
PrepareWorld to be destroyed and deallocated, leaving a dangling pointer
to the underlying Lua state. Usage of this pointer in RunEvents would
cause UB and AccessViolationExceptions.
Removing the call to Stop and placing it at the front of the startup
sequence resolves this problem.