Change UI root to a stack of layers
Prep work for implementing UI objects that may steal input focus, such as popup windows or dropdown menus.

Controls now hold a reference to their parent Control in the hierarchy, the root having a null parent. This is used to check if a control resides on the topmost (active) layer in the UI stack.